Separate out PartialInheritanceMeta.
